#37: further changes in data model, not everything is supposed to be used for moving data between agent and server, i.e. some things can be potentially used from within agent or server (internally)
