+++ model/src/main/scala/za/co/absa/atum/model/MeasureResult.scala
@@ -16,6 +16,4 @@
 
 package za.co.absa.atum.model
 
-case class PartitioningDTO(
-  partitioning: Seq[PartitionDTO]
-)
+trait MeasureResult[T]
